data_utils/requirements.txt
google-cloud-bigquery>=3.10,<4
google-cloud-storage>=2.8,<3
pandas>=2,<3